#28 refactor(diary): 서비스 인터페이스 적용
- 인터페이스, 구현체 분리 작업 진행
